mma: move some methods of PendingRangeChange
This is in preparation for them becoming methods on ExternalRangeChange.
